fixes visualize_morphable_model
