--- menpowidgets/menpo3dwidgets/__init__.py
@@ -1,1 +1,0 @@
-from .base import visualize_morphable_model
